Fix warnings under Xcode 5 and VS 2013 in Poedit and its dependencies, so that
the build is clean.

+++ deps/boost/boost/date_time/date.hpp
@@ -177,7 +177,7 @@ namespace date_time {
       {
         return date_type(date_rep_type(days_) + dd.get_rep());
       }
-      return date_type(date_rep_type(days_) + dd.days());
+      return date_type(date_rep_type(days_) + static_cast<int>(dd.days()));
     }
     date_type operator+=(const duration_type& dd)
     {
